Decoupled business logic from UI functionality
